Move tag actions to contrib/cellTags, menu groups to common
